+++ full-moon/src/ast/versions.rs
@@ -6,6 +6,7 @@ const VERSION_LUA53: u8 = 1 << 2;
 const VERSION_LUA54: u8 = 1 << 3;
 const VERSION_LUAJIT: u8 = 1 << 4;
 const VERSION_CFXLUA: u8 = 1 << 5;
+const VERSION_PLUTO: u8 = 1 << 6;
 
 /// Represents the Lua version(s) to parse as.
 /// Lua 5.1 is always included.
@@ -140,6 +141,25 @@ impl LuaVersion {
     /// Returns true if CFX Lua is enabled.
     pub fn has_cfxlua(self) -> bool {
         cfg!(feature = "cfxlua") && (self.bitfield & VERSION_CFXLUA != 0)
+    }
+
+    /// Creates a new LuaVersion with Pluto, which includes Lua 5.2, 5.3, and Lua 5.4 features.
+    #[cfg(feature = "pluto")]
+    pub fn pluto() -> Self {
+        Self {
+            bitfield: VERSION_LUA52 | VERSION_LUA53 | VERSION_LUA54 | VERSION_PLUTO,
+        }
+    }
+
+    /// Adds Pluto as a version to parse for.
+    #[cfg(feature = "pluto")]
+    pub fn with_pluto(self) -> Self {
+        self | Self::pluto()
+    }
+
+    /// Returns true if Pluto is enabled.
+    pub fn has_pluto(self) -> bool {
+        cfg!(feature = "pluto") && (self.bitfield & VERSION_PLUTO != 0)
     }
 }