+++ full-moon/tests/pluto_cases/pass/ifexpr/source.lua
@@ -1,0 +1,1 @@
+print(if x then "yeah" else "nah" end)