+++ full-moon/tests/pluto_cases/pass/nullcoal/source.lua
@@ -1,0 +1,1 @@
+print(nil ?? "not nil")