+++ full-moon/tests/pluto_cases/pass/unicode_identifier/source.lua
@@ -1,0 +1,2 @@
+local ヴァ = "hello"
+print(ヴァ)