+++ full-moon/tests/pluto_cases/pass/walrus/source.lua
@@ -1,0 +1,8 @@
+if a := "foo" then
+	print(a)
+end
+
+local t = { 1, 2, 3 }
+while val := t:remove(1) do
+	print(val)
+end